Update Undertow web server metrics model to the latest registry-collector separated model
