--- agent/agent-plugins/webserver-tomcat/src/main/java/org/bithon/agent/plugin/tomcat/metric/TomcatWebServerMetricCollector.java
@@ -1,41 +1,0 @@
-/*
- *    Copyright 2020 bithon.org
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-package org.bithon.agent.plugin.tomcat.metric;
-
-import org.apache.tomcat.util.net.AbstractEndpoint;
-import org.bithon.agent.core.metric.domain.web.WebServerMetricRegistry;
-import org.bithon.agent.core.metric.domain.web.WebServerMetrics;
-import org.bithon.agent.core.metric.domain.web.WebServerType;
-
-import java.util.Collections;
-
-/**
- * @author frank.chen021@outlook.com
- * @date 2021/1/13 10:46 下午
- */
-public class TomcatWebServerMetricCollector extends WebServerMetricRegistry {
-
-    public TomcatWebServerMetricCollector(AbstractEndpoint<?> endpoint) {
-        this.register(Collections.singletonList(WebServerType.TOMCAT.type()),
-                      new WebServerMetrics(
-                          endpoint::getConnectionCount,
-                          endpoint::getMaxConnections,
-                          endpoint::getCurrentThreadsBusy,
-                          endpoint::getMaxThreads
-                      ));
-    }
-}
